requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.1.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
pytz==2024.2
typing_extensions==4.12.2
tzlocal==5.2
yarl==1.18.3
